Help from Tina
We changed the experiences datafile to a collection so we could create pages. Then we had to change the default layout to take out the .data so it would read the experience collection.

We also had to make a _config.yml to create a collection in jekyll

Now all info will be pulled from the front matter of each individual .md page for each experience

We could keep the data file configuration if we didn't want individual pages for each experience